+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedLanguageSettingsProvider`
 * [X] 提供基于jsonSchema的（基于JSON指针的）语言设置，用于通过jsonSchema配置语言设置
   * 参见：`icu.windea.ut.toolbox.jsonSchema.JsonSchemaJsonPointerBasedLanguageSettingsProvider`
+* [X] 提供预定义的jsonSchema，用于兼容基于jsonSchema的（基于JSON指针的）语言设置
+  * 参见：`icu.windea.ut.toolbox.jsonSchema.UtJsonSchemaProviderFactory`
 * [X] 提供基于（基于JSON指针的）语言设置的引用解析，以及相关的代码高亮、代码补全、快速文档、代码检查等语言功能
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedReferenceProvider`
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedReferenceUsagesSearcher`
@@ -19,5 +21,5 @@
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedReferenceCompletionProvider`
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedDeclarationDocumentationTargetProvider`
   * 参见：`icu.windea.ut.toolbox.jast.UnresolvedJsonPointerBasedReferenceInspection`
-* [X] 提供预定义的jsonSchema，用于兼容基于jsonSchema的（基于JSON指针的）语言设置
-  * 参见：`icu.windea.ut.toolbox.jsonSchema.UtJsonSchemaProviderFactory`
+* [ ] 提供扩展的折叠规则，适用于基于（基于JSON指针的）语言设置的声明/引用所在的对象或数组
+* [ ] 扩展结构视图，适用于基于（基于JSON指针的）语言设置的声明/引用对应的属性或值